organize selectors
