+++ src/app/ngrx/reducers/selectors/error.type.ts
@@ -1,0 +1,1 @@
+export const errorType = (code) => code === 401 ? 'Authorization' : 'Unknown';
